+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ### Unreleased
 
+- Add a flag to ignore the trailing line break in text height calculations (including tables)
+
 ### [v0.17.1] - 2025-05-02
 
 - Fix null values in table cells rendering as `[object Object]`